+++ HISTORY.md
@@ -1,6 +1,7 @@
 ## 1.0.5 (in progress)
 
 - GATK4: fix option usage for gVCF creation with HaplotypeCaller
+- Add `tools_off: [lumpy-genotype]` option to skip Lumpy genotyping.
 - CWL/WDL: use single file tarballs for complex collections of files like
   aligner, RTG and snpEff indices.